fix(theme): Check for `guide-color-theme` before app load
