+++ guide/.vuepress/templates/index.dev.html
@@ -16,12 +16,22 @@
 	</style>
 	<script>
 		(() => {
+			const { classList } = document.documentElement;
 			const userMode = localStorage.getItem('guide-color-scheme');
 			const systemDarkMode = window.matchMedia && window.matchMedia('(prefers-color-scheme: dark)').matches;
 	
 			if (userMode === 'dark' || (userMode !== 'light' && systemDarkMode)) {
-				document.documentElement.classList.toggle('dark', true);
+				classList.toggle('dark', true);
 			}
+
+			const themes = ['blurple'];
+			const colorTheme = localStorage.getItem('guide-color-theme');
+
+			if (colorTheme !== 'default' && !themes.includes(colorTheme)) {
+				colorTheme = 'default';
+			}
+
+			if (colorTheme !== 'default') classList.add(`guide-theme-${colorTheme}`);
 		})()
 	</script>
 </head>
